+++ nix/packages/nxr.nix
@@ -8,7 +8,7 @@
 
 rustPlatform.buildRustPackage {
   pname = "nxr";
-  version = "3.2.1";
+  version = "3.4.0";
 
   inherit src;